drawline works with floats in uv space instead of pixel coordinates

+++ src/shared/drawing.h
@@ -13,4 +13,4 @@ Some helper functions for drawing 2d images
 
 struct Image;
 
-void DrawLine(Image& image, int x1, int y1, int x2, int y2, const PixelRGBU8& color);
+void DrawLine(Image& image, float x1, float y1, float x2, float y2, const PixelRGBU8& color);
